requirements.txt
fastapi==0.111.1
pywebview==5.1
uvicorn==0.30.3
